+++ lib/python/EasyDel/serve/__init__.py
@@ -1,1 +1,2 @@
-from .serve_utils import shard_params, JAXServer, PyTorchServer,JaxServerConfig
+from .serve_utils import shard_params, JAXServer, PyTorchServer, JaxServerConfig
+
