+++ lib/python/EasyDel/trainer/fsdp_train.py
@@ -9,21 +9,21 @@ from fjutils.easylm import blockwise_cross_entropy, cross_entropy_loss_and_accur
 import wandb
 from datasets import Dataset
 
-from EasyDel.trainer.config import TrainArguments
+from ..trainer.config import TrainArguments
 
 import jax
 import flax
 from transformers import FlaxAutoModelForCausalLM, AutoConfig
 from tqdm import tqdm
-from EasyDel.utils import Timers
-from EasyDel.smi import initialise_tracking, get_mem
+from ..utils import Timers
+from ..smi import initialise_tracking, get_mem
 from jax.experimental.pjit import pjit, with_sharding_constraint
 from jax.sharding import PartitionSpec
 from flax.training import train_state
 from jax import numpy as jnp
 from torch.utils.data import DataLoader
 from fjutils import match_partition_rules, make_shard_and_gather_fns, StreamingCheckpointer, count_params
-from EasyDel.utils import prefix_print
+from ..utils import prefix_print
 
 
 def calculate_accuracy(predictions: jax.Array, targets: jax.Array):
